Codigos no mesmo arquivo que geram problema no contador(uso do init).
